refactor: Edit change textfield color
